+++ get_projects.py
@@ -2,10 +2,11 @@ import asyncio
 import fire
 import json
 import os
-import requests
 
-from oauthlib.oauth2 import BackendApplicationClient
-from requests_oauthlib import OAuth2Session
+from toolbox.get_access_token import get_access_token
+from toolbox.get_operations import get_operations
+from toolbox.get_team_tags import get_team_tags
+from toolbox.post_request import post_request
 
 
 def get_projects(base_url, client_id, client_secret):
@@ -36,44 +37,6 @@ def get_projects(base_url, client_id, client_secret):
         return result
     else:
         print(response)
-
-
-def get_access_token(base_url, client_id, client_secret):
-    client = BackendApplicationClient(client_id=client_id)
-    oauth = OAuth2Session(client=client)
-    token = oauth.fetch_token(
-        token_url=base_url + "/api/oauth/token",
-        client_id=client_id,
-        client_secret=client_secret,
-    )
-    return token["access_token"]
-
-
-def get_operations(file_name):
-    with open(file_name, "r") as file:
-        return json.loads(file.read())
-
-
-def get_team_tags(url, access_token, team_id):
-    operations = get_operations("get_tags.json")
-    operations["variables"]["input"]["teamId"] = team_id
-
-    response = asyncio.run(post_request(url, access_token, operations))
-    if "json" in response.headers["content-type"]:
-        json_response = json.loads(response.text.encode("utf8"))
-        result = json_response["data"]["result"]
-        return result
-    else:
-        print(response)
-
-
-async def post_request(url, access_token, operations):
-    headers = {
-        "Authorization": "Bearer " + access_token,
-        "Content-Type": "application/json",
-    }
-    result = requests.request("POST", url, headers=headers, data=json.dumps(operations))
-    return result
 
 
 if __name__ == "__main__":
