+++ update_project_tags.py
@@ -2,11 +2,12 @@ import asyncio
 import fire
 import json
 import os
-import requests
 import time
 
-from oauthlib.oauth2 import BackendApplicationClient
-from requests_oauthlib import OAuth2Session
+from toolbox.get_access_token import get_access_token
+from toolbox.get_operations import get_operations
+from toolbox.get_team_tags import get_team_tags
+from toolbox.post_request import post_request
 
 
 def update_project_tags(base_url, client_id, client_secret, team_id, method):
@@ -60,36 +61,6 @@ def update_project_tags(base_url, client_id, client_secret, team_id, method):
         print(response)
 
 
-def get_access_token(base_url, client_id, client_secret):
-    client = BackendApplicationClient(client_id=client_id)
-    oauth = OAuth2Session(client=client)
-    token = oauth.fetch_token(
-        token_url=base_url + "/api/oauth/token",
-        client_id=client_id,
-        client_secret=client_secret,
-        verify=False,
-    )
-    return token["access_token"]
-
-
-def get_operations(file_name):
-    with open(file_name, "r") as file:
-        return json.loads(file.read())
-
-
-def get_team_tags(url, access_token, team_id):
-    operations = get_operations("get_tags.json")
-    operations["variables"]["input"]["teamId"] = team_id
-
-    response = asyncio.run(post_request(url, access_token, operations))
-    if "json" in response.headers["content-type"]:
-        json_response = json.loads(response.text.encode("utf8"))
-        result = json_response["data"]["result"]
-        return result
-    else:
-        print(response)
-
-
 def create_tag(url, access_token, team_id, tag_name):
     operations = get_operations("create_tag.json")
     operations["variables"]["input"]["teamId"] = team_id
@@ -115,17 +86,6 @@ def get_single_project(url, access_token, project_id):
         return result
     else:
         print(response)
-
-
-async def post_request(url, access_token, operations):
-    headers = {
-        "Authorization": "Bearer " + access_token,
-        "Content-Type": "application/json",
-    }
-    result = requests.request(
-        "POST", url, headers=headers, data=json.dumps(operations), verify=False
-    )
-    return result
 
 
 if __name__ == "__main__":
